Import raspberry pi debian folder from package source
Follow official raspi distro github:
https://github.com/RPi-Distro/libcamera
We do not use it directly, as this repository import the libcamera
tarball from, therefore loosing all git information on the source code.
We'd rather keep it for our development needs, and get a bit of headache
with the packaging instead. You don't need to say thank you.

Signed-off-by: Benjamin Mugnier <benjamin.mugnier@foss.st.com>